Rename package. Remove unused code.
